optmize code
