+++ pkg/composer/rule_composer.go
@@ -269,13 +269,11 @@ func (r *RuleComposerEngine) Process(req *octollm.Request) (*octollm.Response, e
 			return nil, fmt.Errorf("failed to parse request body: %w", err)
 		}
 		switch body := body.(type) {
-		case *openai.ApiChatCompletionsRequest:
+		case *openai.ChatCompletionRequest:
 			r.Model = body.Model
-		// case *openai.ChatCompletionNewParams:
-		// 	r.Model = body.Model
 		case *openaiSDK.ChatCompletionNewParams:
 			r.Model = body.Model
-		case *anthropic.ApiMessagesRequest:
+		case *anthropic.ClaudeMessagesRequest:
 			r.Model = body.Model
 		case *anthropicSDK.MessageNewParams:
 			r.Model = string(body.Model)
