+++ pkg/engines/converter/claude.go
@@ -6,28 +6,25 @@ import (
 	"errors"
 	"fmt"
 
-	anthropicSDK "github.com/anthropics/anthropic-sdk-go"
-	openaiSDK "github.com/openai/openai-go/v3"
 	"github.com/sirupsen/logrus"
-	"github.com/tidwall/gjson"
 
 	"github.com/infinigence/octollm/pkg/octollm"
 	"github.com/infinigence/octollm/pkg/types/anthropic"
 	"github.com/infinigence/octollm/pkg/types/openai"
 )
 
-// ChatCompletionToClaudeMessages is an engine that handles ClaudeMessages requests with an underlying ChatCompletions engine.
-type ChatCompletionsToClaudeMessages struct {
+type ChatCompletionToClaudeMessages struct {
 	next octollm.Engine // the engine that can handle ChatCompletions requests
 }
 
-var _ octollm.Engine = (*ChatCompletionsToClaudeMessages)(nil)
+var _ octollm.Engine = (*ChatCompletionToClaudeMessages)(nil)
 
-func NewChatCompletionsToClaudeMessages(next octollm.Engine) *ChatCompletionsToClaudeMessages {
-	return &ChatCompletionsToClaudeMessages{next: next}
+func NewChatCompletionToClaudeMessages(next octollm.Engine) *ChatCompletionToClaudeMessages {
+	return &ChatCompletionToClaudeMessages{next: next}
 }
 
-func (e *ChatCompletionsToClaudeMessages) Process(req *octollm.Request) (*octollm.Response, error) {
+func (e *ChatCompletionToClaudeMessages) Process(req *octollm.Request) (*octollm.Response, error) {
+	logrus.WithContext(req.Context()).Infof("converting request body to ChatCompletions format from ClaudeMessages")
 	newBody, err := e.convertRequestBody(req.Context(), req.Body)
 	if err != nil {
 		return nil, fmt.Errorf("failed to convert request body: %w", err)
@@ -35,13 +32,13 @@ func (e *ChatCompletionsToClaudeMessages) Process(req *octollm.Request) (*octoll
 	req.Format = octollm.APIFormatChatCompletions
 	req.Body = newBody
 
-	// 4. Call Next Engine
+	// Call Next Engine
 	resp, err := e.next.Process(req)
 	if err != nil {
 		return nil, err
 	}
 
-	// 5. Convert Response
+	// Convert Response
 	if resp.Stream != nil {
 		newStream, err := e.convertStreamResponse(req.Context(), resp.Stream)
 		if err != nil {
@@ -59,215 +56,323 @@ func (e *ChatCompletionsToClaudeMessages) Process(req *octollm.Request) (*octoll
 	return resp, nil
 }
 
-func (e *ChatCompletionsToClaudeMessages) convertRequestBody(ctx context.Context, srcBody *octollm.UnifiedBody) (*octollm.UnifiedBody, error) {
+func (e *ChatCompletionToClaudeMessages) convertRequestBody(ctx context.Context, srcBody *octollm.UnifiedBody) (*octollm.UnifiedBody, error) {
 	// Parse Input as Anthropic Request
 	anthropicReq, err := srcBody.Parsed()
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse request body: %w", err)
 	}
 
-	src, ok := anthropicReq.(*anthropic.MessageNewParams)
+	src, ok := anthropicReq.(*anthropic.ClaudeMessagesRequest)
 	if !ok {
-		// Fallback for when parsed type isn't what we expect
-		return nil, fmt.Errorf("parsed body is not *anthropic.MessageNewParams, got %T", anthropicReq)
+		return nil, fmt.Errorf("parsed body is not *anthropic.ClaudeMessagesRequest, got %T", anthropicReq)
 	}
 
-	dst := &openai.ChatCompletionNewParams{}
+	dst := &openai.ChatCompletionRequest{}
 
-	if src.Stream.Valid() {
-		dst.Stream = openaiSDK.Bool(src.Stream.Value)
+	// Stream
+	if src.Stream != nil {
+		dst.Stream = src.Stream
 	}
 
 	// Model
-	dst.Model = string(src.Model)
+	dst.Model = src.Model
 
 	// MaxTokens
-	dst.MaxTokens = openaiSDK.Int(src.MaxTokens)
+	if src.MaxTokens > 0 {
+		maxTokens := int(src.MaxTokens)
+		dst.MaxTokens = &maxTokens
+	}
 
 	// Temperature
-	if src.Temperature.Valid() {
-		dst.Temperature = openaiSDK.Float(src.Temperature.Value)
+	if src.Temperature != nil {
+		dst.Temperature = src.Temperature
 	}
 
 	// TopP
-	if src.TopP.Valid() {
-		dst.TopP = openaiSDK.Float(src.TopP.Value)
+	if src.TopP != nil {
+		dst.TopP = src.TopP
 	}
 
-	// Stop Sequences - Skipped for now due to complex union type mapping
-	// if len(src.StopSequences) > 0 { ... }
-
-	// System Prompt -> Messages
-	var messages []openaiSDK.ChatCompletionMessageParamUnion
-	for _, sysBlock := range src.System {
-		// sysBlock is anthropic.TextBlockParam
-		// Text field is string.
-		// System Message
-		msg := openaiSDK.SystemMessage(sysBlock.Text)
-		messages = append(messages, msg)
+	// TopK
+	if src.TopK != nil {
+		topK := int(*src.TopK)
+		dst.TopK = &topK
 	}
 
-	// Messages
+	// Stop Sequences
+	if len(src.StopSequences) > 0 {
+		dst.Stop = src.StopSequences
+	}
+
+	// Convert System Prompt to Messages
+	var messages []*openai.Message
+	if src.System != nil {
+		switch sys := src.System.(type) {
+		case anthropic.SystemString:
+			messages = append(messages, &openai.Message{
+				Role:    anthropic.MessageParamRoleSystem,
+				Content: openai.MessageContentString(sys),
+			})
+		case anthropic.SystemBlocks:
+			for _, block := range sys {
+				messages = append(messages, &openai.Message{
+					Role:    anthropic.MessageParamRoleSystem,
+					Content: openai.MessageContentString(block.Text),
+				})
+			}
+		}
+	}
+
+	// Convert Messages
 	for _, msg := range src.Messages {
-		role := msg.Role // anthropic.MessageParamRole
-
-		if role == anthropicSDK.MessageParamRoleUser {
-			var contentParts []openaiSDK.ChatCompletionContentPartUnionParam
+		role := msg.Role
+		if role == anthropic.MessageParamRoleUser {
+			var contentParts []*openai.MessageContentItem
 			for _, block := range msg.Content {
-				// block is anthropic.ContentBlockParamUnion
-				if block.OfText != nil {
-					// Text
-					contentParts = append(contentParts, openaiSDK.TextContentPart(block.OfText.Text))
-				} else if block.OfImage != nil {
-					// Image
-					src := block.OfImage.Source
-					var url string
-					if src.OfBase64 != nil {
-						mediaType := src.OfBase64.MediaType
-						data := src.OfBase64.Data
-						url = fmt.Sprintf("data:%s;base64,%s", mediaType, data)
-					} else if src.OfURL != nil {
-						url = src.OfURL.URL
+				contentBlock, ok := block.(*anthropic.MessageContentBlock)
+				if !ok {
+					// Handle MessageContentString case
+					if str, ok := block.(anthropic.MessageContentString); ok {
+						contentParts = append(contentParts, &openai.MessageContentItem{
+							Type: anthropic.MessageContentTextType,
+							Text: string(str),
+						})
 					}
-					part := openaiSDK.ChatCompletionContentPartImageImageURLParam{
-						URL: url,
-					}
-					contentParts = append(contentParts, openaiSDK.ImageContentPart(part))
-				} else if block.OfToolResult != nil {
-					// Tool Result
-					var contentParts []openaiSDK.ChatCompletionContentPartTextParam
-					for _, content := range block.OfToolResult.Content {
-						if content.OfText == nil {
-							return nil, fmt.Errorf("tool result content is not text")
-						}
-						part := openaiSDK.ChatCompletionContentPartTextParam{
-							Text: content.OfText.Text,
-						}
-						contentParts = append(contentParts, part)
-					}
-					messages = append(messages, openaiSDK.ToolMessage(contentParts, block.OfToolResult.ToolUseID))
+					continue
 				}
-				// TODO: other block types
+
+				switch contentBlock.Type {
+				case anthropic.MessageContentTextType:
+					if contentBlock.Text != nil {
+						contentParts = append(contentParts, &openai.MessageContentItem{
+							Type: anthropic.MessageContentTextType,
+							Text: *contentBlock.Text,
+						})
+					}
+				case anthropic.MessageContentImageType:
+					if contentBlock.Source != nil {
+						url := ""
+						if contentBlock.Source.Type == "base64" && contentBlock.Source.Data != nil {
+							mediaType := "image/jpeg"
+							if contentBlock.Source.MediaType != "" {
+								mediaType = contentBlock.Source.MediaType
+							}
+							dataStr := fmt.Sprintf("%v", contentBlock.Source.Data)
+							url = fmt.Sprintf("data:%s;base64,%s", mediaType, dataStr)
+						} else if contentBlock.Source.Type == "url" {
+							url = contentBlock.Source.Url
+						}
+						contentParts = append(contentParts, &openai.MessageContentItem{
+							Type: "image_url",
+							ImageURL: &openai.MessageContentItemImageURL{
+								URL: url,
+							},
+						})
+					}
+				case anthropic.MessageContentToolResultType:
+					// Tool result should be converted to tool message
+					if contentBlock.MessageContentToolResult != nil && contentBlock.ToolUseID != nil {
+						var resultText string
+						if len(contentBlock.MessageContentToolResult.Content) > 0 {
+							for _, c := range contentBlock.MessageContentToolResult.Content {
+								resultText += c.ExtractText()
+							}
+						}
+						messages = append(messages, &openai.Message{
+							Role:       anthropic.MessageParamRoleTool,
+							Content:    openai.MessageContentString(resultText),
+							ToolCallID: *contentBlock.ToolUseID,
+						})
+					}
+				}
 			}
 
 			if len(contentParts) > 0 {
-				messages = append(messages, openaiSDK.UserMessage(contentParts))
+				messages = append(messages, &openai.Message{
+					Role:    anthropic.MessageParamRoleUser,
+					Content: openai.MessageContentArray(contentParts),
+				})
 			}
 
-		} else if role == anthropicSDK.MessageParamRoleAssistant {
-			var contentParts []openaiSDK.ChatCompletionAssistantMessageParamContentArrayOfContentPartUnion
-			var toolCalls []openaiSDK.ChatCompletionMessageToolCallUnionParam
+		} else if role == anthropic.MessageParamRoleAssistant {
+			var contentParts []*openai.MessageContentItem
+			var toolCalls []*openai.ToolCall
+			toolCallIndex := 0
+
 			for _, block := range msg.Content {
-				// block is anthropic.ContentBlockParamUnion
-				if block.OfText != nil {
-					// Text
-					part := openaiSDK.ChatCompletionContentPartTextParam{
-						Text: block.OfText.Text,
+				contentBlock, ok := block.(*anthropic.MessageContentBlock)
+				if !ok {
+					// Handle MessageContentString case
+					if str, ok := block.(anthropic.MessageContentString); ok {
+						contentParts = append(contentParts, &openai.MessageContentItem{
+							Type: anthropic.MessageContentTextType,
+							Text: string(str),
+						})
 					}
-					contentParts = append(contentParts, openaiSDK.ChatCompletionAssistantMessageParamContentArrayOfContentPartUnion{OfText: &part})
-				} else if block.OfToolUse != nil {
-					inputs, err := json.Marshal(block.OfToolUse.Input)
+					continue
+				}
+
+				if contentBlock.Type == "text" && contentBlock.Text != nil {
+					contentParts = append(contentParts, &openai.MessageContentItem{
+						Type: anthropic.MessageContentTextType,
+						Text: *contentBlock.Text,
+					})
+				} else if contentBlock.Type == "tool_use" && contentBlock.MessageContentToolUse != nil {
+					inputs, err := json.Marshal(contentBlock.MessageContentToolUse.Input)
 					if err != nil {
 						return nil, fmt.Errorf("failed to marshal tool use input: %w", err)
 					}
-					// Tool Use
-					toolCall := &openaiSDK.ChatCompletionMessageFunctionToolCallParam{
-						ID: block.OfToolUse.ID,
-						Function: openaiSDK.ChatCompletionMessageFunctionToolCallFunctionParam{
-							Name:      block.OfToolUse.Name,
+					toolCalls = append(toolCalls, &openai.ToolCall{
+						ID:    contentBlock.MessageContentToolUse.ID,
+						Index: toolCallIndex,
+						Type:  "function",
+						Function: &openai.ToolCallFunction{
+							Name:      contentBlock.MessageContentToolUse.Name,
 							Arguments: string(inputs),
 						},
-					}
-					toolCalls = append(toolCalls, openaiSDK.ChatCompletionMessageToolCallUnionParam{OfFunction: toolCall})
+					})
+					toolCallIndex++
 				}
-				// Assistant messages do not support images in OpenAI
 			}
 
-			assistantMsg := openaiSDK.AssistantMessage(contentParts)
+			assistantMsg := &openai.Message{
+				Role: anthropic.MessageParamRoleAssistant,
+			}
+			if len(contentParts) > 0 {
+				assistantMsg.Content = openai.MessageContentArray(contentParts)
+			}
 			if len(toolCalls) > 0 {
-				assistantMsg.OfAssistant.ToolCalls = toolCalls
+				assistantMsg.ToolCalls = toolCalls
 			}
 			messages = append(messages, assistantMsg)
 		}
 	}
 	dst.Messages = messages
 
-	// Tools
+	// Convert Tools
 	for _, tool := range src.Tools {
-		if tool.OfTool == nil {
+		if tool.Name == "" {
 			continue
 		}
-		// Function
-		fdp := openaiSDK.FunctionDefinitionParam{
-			Name:        tool.OfTool.Name,
-			Description: openaiSDK.String(tool.OfTool.Description.Value),
-			Parameters: openaiSDK.FunctionParameters{
-				"type":       tool.OfTool.InputSchema.Type,
-				"properties": tool.OfTool.InputSchema.Properties,
-				"required":   tool.OfTool.InputSchema.Required,
-			},
+		var params json.RawMessage
+		if tool.InputSchema != nil {
+			if raw, ok := tool.InputSchema.(json.RawMessage); ok {
+				params = raw
+			} else {
+				// Try to marshal it
+				paramsBytes, err := json.Marshal(tool.InputSchema)
+				if err == nil {
+					params = paramsBytes
+				}
+			}
 		}
-		dst.Tools = append(dst.Tools, openaiSDK.ChatCompletionFunctionTool(fdp))
+		dst.Tools = append(dst.Tools, &openai.Tool{
+			Type: "function",
+			Function: openai.ToolFunction{
+				Name:        &tool.Name,
+				Description: &tool.Description,
+				Parameters:  params,
+			},
+		})
+	}
+
+	// Convert ToolChoice
+	if src.ToolChoice != nil {
+		switch src.ToolChoice.Type {
+		case "auto":
+			auto := "auto"
+			dst.ToolChoice = &openai.ToolChoice{String: &auto}
+		case "any":
+			required := "required"
+			dst.ToolChoice = &openai.ToolChoice{String: &required}
+		case "tool":
+			if src.ToolChoice.Name != nil {
+				dst.ToolChoice = &openai.ToolChoice{
+					Function: &openai.ToolChoiceFunction{
+						Type: "function",
+						Function: &openai.ToolChoiceFunctionParam{
+							Name: *src.ToolChoice.Name,
+						},
+					},
+				}
+			}
+		}
 	}
 
 	// Convert to UnifiedBody
-	newBody := octollm.NewBodyFromBytes([]byte{}, &octollm.JSONParser[openai.ChatCompletionNewParams]{})
+	newBody := octollm.NewBodyFromBytes([]byte{}, &octollm.JSONParser[openai.ChatCompletionRequest]{})
 	newBody.SetParsed(dst)
 
 	return newBody, nil
 }
 
-func (e *ChatCompletionsToClaudeMessages) convertNonStreamResponseBody(ctx context.Context, srcBody *octollm.UnifiedBody) (*octollm.UnifiedBody, error) {
+func (e *ChatCompletionToClaudeMessages) convertNonStreamResponseBody(ctx context.Context, srcBody *octollm.UnifiedBody) (*octollm.UnifiedBody, error) {
 	// Parse Input as OpenAI Response
 	parsed, err := srcBody.Parsed()
 	if err != nil {
-		return nil, fmt.Errorf("failed to parse request body: %w", err)
+		return nil, fmt.Errorf("failed to parse response body: %w", err)
 	}
 
-	openaiResp, ok := parsed.(*openaiSDK.ChatCompletion)
+	openaiResp, ok := parsed.(*openai.ChatCompletionResponse)
 	if !ok {
-		return nil, fmt.Errorf("parsed body is not *openaiSDK.ChatCompletion, got %T", parsed)
+		return nil, fmt.Errorf("parsed body is not *openai.ChatCompletionResponse, got %T", parsed)
 	}
 
 	// Construct Claude Response
-	claudeResp := &anthropic.MessageSimple{
-		ID:    openaiResp.ID,
+	claudeResp := &anthropic.ClaudeMessagesResponse{
+		ID:    openaiResp.Id,
 		Type:  "message",
 		Role:  "assistant",
 		Model: openaiResp.Model,
-		Usage: &anthropic.MessageUsage{
-			InputTokens:  int(openaiResp.Usage.PromptTokens),
-			OutputTokens: int(openaiResp.Usage.CompletionTokens),
-		},
+	}
+
+	// Usage
+	if openaiResp.Usage != nil {
+		claudeResp.Usage = &anthropic.Usage{
+			InputTokens:  int64(openaiResp.Usage.PromptTokens),
+			OutputTokens: int64(openaiResp.Usage.CompletionTokens),
+		}
+		// Handle cached tokens if available
+		if openaiResp.Usage.PromptTokensDetails != nil && openaiResp.Usage.PromptTokensDetails.CachedTokens > 0 {
+			cached := int64(openaiResp.Usage.PromptTokensDetails.CachedTokens)
+			claudeResp.Usage.CacheReadInputTokens = &cached
+		}
 	}
 
 	// Choices
 	if len(openaiResp.Choices) > 0 {
 		choice := openaiResp.Choices[0]
 		// Finish Reason
-		fr := string(choice.FinishReason)
-		mappedFr := e.mapFinishReason(fr)
-		claudeResp.StopReason = &mappedFr
+		claudeResp.StopReason = e.mapFinishReason(choice.FinishReason)
 
 		msg := choice.Message
-		// Content
-		if msg.JSON.Content.Valid() {
-			claudeResp.Content = append(claudeResp.Content, anthropic.ContentBlockText{
-				Type: "text",
-				Text: msg.Content,
-			})
-		}
-		for _, toolCall := range msg.ToolCalls {
-			if !gjson.Valid(toolCall.Function.Arguments) {
-				logrus.WithContext(ctx).Warnf("invalid tool call arguments: %s", toolCall.Function.Arguments)
-				continue
+		if msg != nil {
+			// Content
+			if msg.Content != nil {
+				text := msg.Content.ExtractText()
+				if text != "" {
+					claudeResp.Content = append(claudeResp.Content, &anthropic.MessageContentBlock{
+						Type: "text",
+						Text: &text,
+					})
+				}
 			}
-			claudeResp.Content = append(claudeResp.Content, anthropic.ContentBlockToolUse{
-				Type:  "tool_use",
-				ID:    toolCall.ID,
-				Name:  toolCall.Function.Name,
-				Input: json.RawMessage(toolCall.Function.Arguments),
-			})
+
+			// Tool Calls
+			for _, toolCall := range msg.ToolCalls {
+				if toolCall.Function == nil {
+					continue
+				}
+				claudeResp.Content = append(claudeResp.Content, &anthropic.MessageContentBlock{
+					Type: "tool_use",
+					MessageContentToolUse: &anthropic.MessageContentToolUse{
+						ID:    toolCall.ID,
+						Name:  toolCall.Function.Name,
+						Input: json.RawMessage(toolCall.Function.Arguments),
+					},
+				})
+			}
 		}
 	}
 
@@ -277,12 +382,12 @@ func (e *ChatCompletionsToClaudeMessages) convertNonStreamResponseBody(ctx conte
 		return nil, fmt.Errorf("failed to marshal claude response: %w", err)
 	}
 
-	newBody := octollm.NewBodyFromBytes(claudeBytes, &octollm.JSONParser[anthropicSDK.Message]{})
+	newBody := octollm.NewBodyFromBytes(claudeBytes, &octollm.JSONParser[anthropic.ClaudeMessagesResponse]{})
 
 	return newBody, nil
 }
 
-func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Context, src *octollm.StreamChan) (*octollm.StreamChan, error) {
+func (e *ChatCompletionToClaudeMessages) convertStreamResponse(ctx context.Context, src *octollm.StreamChan) (*octollm.StreamChan, error) {
 	inCh := src.Chan()
 	outCh := make(chan *octollm.StreamChunk)
 
@@ -305,10 +410,10 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 			blockTypeTool
 		)
 		currentBlockType := blockTypeNone
-		currentToolCallIndex := int64(-1) // Track which OpenAI tool call index is in current block
+		currentToolCallIndex := -1 // Track which OpenAI tool call index is in current block
 
 		var pendingFinishReason *string
-		var pendingUsage *openaiSDK.CompletionUsage
+		var pendingUsage *openai.Usage
 
 		for chunk := range inCh {
 			if ctx.Err() != nil {
@@ -327,7 +432,7 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 
 				// Send content_block_stop for current block if one exists
 				if currentBlockType != blockTypeNone {
-					blockStop := &anthropic.MessageStreamEvent{
+					blockStop := &anthropic.ClaudeMessagesStreamEvent{
 						Type:  "content_block_stop",
 						Index: intPtr(currentBlockIndex),
 					}
@@ -341,16 +446,19 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 				// When we have both finish_reason and usage, send message_delta and message_stop
 				if pendingFinishReason != nil {
 					mappedFr := e.mapFinishReason(*pendingFinishReason)
-					msgDelta := &anthropic.MessageStreamEvent{
+					msgDelta := &anthropic.ClaudeMessagesStreamEvent{
 						Type: "message_delta",
-						Delta: &anthropic.MessageDelta{
-							StopReason: &mappedFr,
-						},
 					}
+					deltaData := &anthropic.ApiMessageDelta{
+						StopReason: &mappedFr,
+					}
+					deltaBytes, _ := json.Marshal(deltaData)
+					msgDelta.DeltaRaw = deltaBytes
+
 					if pendingUsage != nil {
-						msgDelta.Usage = &anthropic.MessageUsage{
-							InputTokens:  int(pendingUsage.PromptTokens),
-							OutputTokens: int(pendingUsage.CompletionTokens),
+						msgDelta.Usage = &anthropic.Usage{
+							InputTokens:  int64(pendingUsage.PromptTokens),
+							OutputTokens: int64(pendingUsage.CompletionTokens),
 						}
 					}
 					if err := e.sendEvent(outCh, msgDelta); err != nil {
@@ -361,7 +469,7 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 				}
 
 				// Send message_stop
-				msgStop := &anthropic.MessageStreamEvent{
+				msgStop := &anthropic.ClaudeMessagesStreamEvent{
 					Type: "message_stop",
 				}
 				if err := e.sendEvent(outCh, msgStop); err != nil {
@@ -370,26 +478,25 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 				break
 			}
 
-			// Assert to *openaiSDK.ChatCompletionChunk
-			openaiChunk, ok := parsed.(*openaiSDK.ChatCompletionChunk)
+			openaiChunk, ok := parsed.(*openai.ChatCompletionResponse)
 			if !ok {
-				logrus.WithContext(ctx).Errorf("parsed stream chunk is not *openai.ChatCompletionChunk, got %T", parsed)
+				logrus.WithContext(ctx).Errorf("parsed stream chunk is not *openai.ChatCompletionStreamResponse, got %T", parsed)
 				continue
 			}
 
 			if !started {
-				msgID = openaiChunk.ID
+				msgID = openaiChunk.Id
 				model = openaiChunk.Model
 				// Send message_start
-				msgStart := &anthropic.MessageStreamEvent{
+				msgStart := &anthropic.ClaudeMessagesStreamEvent{
 					Type: "message_start",
-					Message: &anthropic.MessageSimple{
+					Message: &anthropic.ClaudeMessagesResponse{
 						ID:      msgID,
 						Type:    "message",
 						Role:    "assistant",
 						Model:   model,
-						Content: []anthropic.ContentBlock{},
-						Usage:   &anthropic.MessageUsage{InputTokens: 0, OutputTokens: 0}, // Placeholder
+						Content: []anthropic.MessageContent{},
+						Usage:   &anthropic.Usage{InputTokens: 0, OutputTokens: 0}, // Placeholder
 					},
 				}
 				if err := e.sendEvent(outCh, msgStart); err != nil {
@@ -401,21 +508,24 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 
 			// Extract Delta
 			var deltaContent string
-			var toolCalls []openaiSDK.ChatCompletionChunkChoiceDeltaToolCall
+			var toolCalls []*openai.ToolCall
 
 			if len(openaiChunk.Choices) > 0 {
 				choice := openaiChunk.Choices[0]
-				deltaContent = choice.Delta.Content
-				toolCalls = choice.Delta.ToolCalls
+				if choice.Delta != nil {
+					if choice.Delta.Content != nil {
+						deltaContent = choice.Delta.Content.ExtractText()
+					}
+					toolCalls = choice.Delta.ToolCalls
+				}
 				if finishReason := choice.FinishReason; finishReason != "" {
-					// record first finish reason
-					pendingFinishReason = &finishReason
+					pendingFinishReason = &choice.FinishReason
 				}
 			}
 
 			// Check if this chunk has usage info
-			if openaiChunk.JSON.Usage.Valid() && openaiChunk.Usage.JSON.PromptTokens.Valid() {
-				pendingUsage = &openaiChunk.Usage
+			if openaiChunk.Usage != nil {
+				pendingUsage = openaiChunk.Usage
 			}
 
 			// Handle text content
@@ -424,18 +534,15 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 				needNewBlock := false
 				switch currentBlockType {
 				case blockTypeNone:
-					// No current block, need to start one
 					needNewBlock = true
 				case blockTypeTool:
-					// Switching from tool to text, need new block
 					needNewBlock = true
 				}
-				// If currentBlockType == blockTypeText, continue with existing text block
 
 				if needNewBlock {
 					// Close previous block if exists
 					if currentBlockType != blockTypeNone {
-						blockStop := &anthropic.MessageStreamEvent{
+						blockStop := &anthropic.ClaudeMessagesStreamEvent{
 							Type:  "content_block_stop",
 							Index: intPtr(currentBlockIndex),
 						}
@@ -449,12 +556,13 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 					currentBlockIndex++
 
 					// Start text block
-					blockStart := &anthropic.MessageStreamEvent{
+					emptyText := ""
+					blockStart := &anthropic.ClaudeMessagesStreamEvent{
 						Type:  "content_block_start",
 						Index: intPtr(currentBlockIndex),
-						ContentBlock: &anthropic.ContentBlockText{
+						ContentBlock: &anthropic.MessageContentBlock{
 							Type: "text",
-							Text: "",
+							Text: &emptyText,
 						},
 					}
 					if err := e.sendEvent(outCh, blockStart); err != nil {
@@ -465,14 +573,17 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 				}
 
 				// Send text delta
-				deltaEvent := &anthropic.MessageStreamEvent{
+				deltaEvent := &anthropic.ClaudeMessagesStreamEvent{
 					Type:  "content_block_delta",
 					Index: intPtr(currentBlockIndex),
-					Delta: &anthropic.ContentBlockTextDelta{
-						Type: "text_delta",
-						Text: deltaContent,
-					},
 				}
+				deltaData := &anthropic.ApiContentBlockDelta{
+					Type: "text_delta",
+					Text: &deltaContent,
+				}
+				deltaBytes, _ := json.Marshal(deltaData)
+				deltaEvent.DeltaRaw = deltaBytes
+
 				if err := e.sendEvent(outCh, deltaEvent); err != nil {
 					logrus.WithContext(ctx).Errorf("failed to send content_block_delta event: %v", err)
 					continue
@@ -486,14 +597,11 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 					needNewBlock := false
 					switch currentBlockType {
 					case blockTypeNone:
-						// No current block, need to start one
 						needNewBlock = true
 					case blockTypeText:
-						// Switching from text to tool, need new block
 						needNewBlock = true
 					case blockTypeTool:
 						if currentToolCallIndex != toolCall.Index {
-							// Switching to different tool call, need new block
 							needNewBlock = true
 						}
 					}
@@ -501,7 +609,7 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 					if needNewBlock {
 						// Close previous block if exists
 						if currentBlockType != blockTypeNone {
-							blockStop := &anthropic.MessageStreamEvent{
+							blockStop := &anthropic.ClaudeMessagesStreamEvent{
 								Type:  "content_block_stop",
 								Index: intPtr(currentBlockIndex),
 							}
@@ -515,14 +623,16 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 						currentBlockIndex++
 
 						// Start tool_use block
-						blockStart := &anthropic.MessageStreamEvent{
+						blockStart := &anthropic.ClaudeMessagesStreamEvent{
 							Type:  "content_block_start",
 							Index: intPtr(currentBlockIndex),
-							ContentBlock: &anthropic.ContentBlockToolUse{
-								Type:  "tool_use",
-								ID:    toolCall.ID,
-								Name:  toolCall.Function.Name,
-								Input: json.RawMessage("{}"),
+							ContentBlock: &anthropic.MessageContentBlock{
+								Type: "tool_use",
+								MessageContentToolUse: &anthropic.MessageContentToolUse{
+									ID:    toolCall.ID,
+									Name:  toolCall.Function.Name,
+									Input: json.RawMessage("{}"),
+								},
 							},
 						}
 						if err := e.sendEvent(outCh, blockStart); err != nil {
@@ -534,15 +644,18 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 					}
 
 					// Send input_json_delta for tool call arguments
-					if toolCall.Function.Arguments != "" {
-						deltaEvent := &anthropic.MessageStreamEvent{
+					if toolCall.Function != nil && toolCall.Function.Arguments != "" {
+						deltaEvent := &anthropic.ClaudeMessagesStreamEvent{
 							Type:  "content_block_delta",
 							Index: intPtr(currentBlockIndex),
-							Delta: &anthropic.ContentBlockInputJSONDelta{
-								Type:        "input_json_delta",
-								PartialJSON: toolCall.Function.Arguments,
-							},
 						}
+						deltaData := &anthropic.ApiContentBlockDelta{
+							Type:        "input_json_delta",
+							PartialJSON: &toolCall.Function.Arguments,
+						}
+						deltaBytes, _ := json.Marshal(deltaData)
+						deltaEvent.DeltaRaw = deltaBytes
+
 						if err := e.sendEvent(outCh, deltaEvent); err != nil {
 							logrus.WithContext(ctx).Errorf("failed to send input_json_delta event: %v", err)
 							continue
@@ -557,17 +670,17 @@ func (e *ChatCompletionsToClaudeMessages) convertStreamResponse(ctx context.Cont
 	return newStream, nil
 }
 
-func (e *ChatCompletionsToClaudeMessages) sendEvent(ch chan<- *octollm.StreamChunk, event *anthropic.MessageStreamEvent) error {
+func (e *ChatCompletionToClaudeMessages) sendEvent(ch chan<- *octollm.StreamChunk, event *anthropic.ClaudeMessagesStreamEvent) error {
 	bytes, err := json.Marshal(event)
 	if err != nil {
 		return fmt.Errorf("failed to marshal claude stream event: %w", err)
 	}
-	body := octollm.NewBodyFromBytes(bytes, &octollm.JSONParser[anthropicSDK.BetaRawMessageStreamEventUnion]{})
+	body := octollm.NewBodyFromBytes(bytes, &octollm.JSONParser[anthropic.ClaudeMessagesStreamEvent]{})
 	ch <- &octollm.StreamChunk{Body: body, Metadata: map[string]string{"event": event.Type}}
 	return nil
 }
 
-func (e *ChatCompletionsToClaudeMessages) mapFinishReason(fr string) string {
+func (e *ChatCompletionToClaudeMessages) mapFinishReason(fr string) string {
 	switch fr {
 	case "stop":
 		return "end_turn"
