+++ pkg/types/anthropic/messages_response.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 )
 
-// ApiMessagesResponse represents a complete Anthropic Messages API response
-type ApiMessagesResponse struct {
+// ClaudeMessagesResponse represents a complete Anthropic Messages API response
+type ClaudeMessagesResponse struct {
 	ID           string           `json:"id"`
 	Type         string           `json:"type"`
 	Role         string           `json:"role"`
@@ -16,9 +16,9 @@ type ApiMessagesResponse struct {
 	Usage        *Usage           `json:"usage"`
 }
 
-// UnmarshalJSON implements custom JSON unmarshaling for ApiMessagesResponse
-func (r *ApiMessagesResponse) UnmarshalJSON(data []byte) error {
-	type Alias ApiMessagesResponse
+// UnmarshalJSON implements custom JSON unmarshaling for ClaudeMessagesResponse
+func (r *ClaudeMessagesResponse) UnmarshalJSON(data []byte) error {
+	type Alias ClaudeMessagesResponse
 	aux := &struct {
 		Content json.RawMessage `json:"content"`
 		*Alias
@@ -60,15 +60,15 @@ type Usage struct {
 	CacheReadInputTokens *int64 `json:"cache_read_input_tokens,omitempty"`
 }
 
-// ApiMessagesStreamEvent represents a streaming event
+// ClaudeMessagesStreamEvent represents a streaming event
 // Aligned with MessageStreamEventUnion from SDK
-type ApiMessagesStreamEvent struct {
+type ClaudeMessagesStreamEvent struct {
 	// Event type: "message_start", "content_block_start", "content_block_delta",
 	// "content_block_stop", "message_delta", "message_stop", "ping", "error"
 	Type string `json:"type"`
 
 	// For message_start event
-	Message *ApiMessagesResponse `json:"message,omitempty"`
+	Message *ClaudeMessagesResponse `json:"message,omitempty"`
 
 	// For content_block_start event
 	Index *int `json:"index,omitempty"`
@@ -80,7 +80,7 @@ type ApiMessagesStreamEvent struct {
 }
 
 // GetContentBlockDelta returns the delta as ApiContentBlockDelta if applicable
-func (e *ApiMessagesStreamEvent) GetContentBlockDelta() (*ApiContentBlockDelta, error) {
+func (e *ClaudeMessagesStreamEvent) GetContentBlockDelta() (*ApiContentBlockDelta, error) {
 	if e.Type != "content_block_delta" || len(e.DeltaRaw) == 0 {
 		return nil, nil
 	}
@@ -92,7 +92,7 @@ func (e *ApiMessagesStreamEvent) GetContentBlockDelta() (*ApiContentBlockDelta, 
 }
 
 // GetMessageDelta returns the delta as ApiMessageDelta if applicable
-func (e *ApiMessagesStreamEvent) GetMessageDelta() (*ApiMessageDelta, error) {
+func (e *ClaudeMessagesStreamEvent) GetMessageDelta() (*ApiMessageDelta, error) {
 	if e.Type != "message_delta" || len(e.DeltaRaw) == 0 {
 		return nil, nil
 	}
@@ -130,7 +130,7 @@ type APIError struct {
 }
 
 // ExtractText extracts all text content from the response
-func (r *ApiMessagesResponse) ExtractText() string {
+func (r *ClaudeMessagesResponse) ExtractText() string {
 	text := ""
 	for _, block := range r.Content {
 		if blockBlock, ok := block.(*MessageContentBlock); ok {
@@ -148,7 +148,7 @@ func (r *ApiMessagesResponse) ExtractText() string {
 }
 
 // ExtractToolUses extracts all tool use blocks from the response
-func (r *ApiMessagesResponse) ExtractToolUses() []*MessageContentToolUse {
+func (r *ClaudeMessagesResponse) ExtractToolUses() []*MessageContentToolUse {
 	var toolUses []*MessageContentToolUse
 	for _, block := range r.Content {
 		if blockBlock, ok := block.(*MessageContentBlock); ok && blockBlock.Type == "tool_use" && blockBlock.MessageContentToolUse != nil {
@@ -159,26 +159,26 @@ func (r *ApiMessagesResponse) ExtractToolUses() []*MessageContentToolUse {
 }
 
 // IsToolUse checks if the response contains tool use
-func (r *ApiMessagesResponse) IsToolUse() bool {
+func (r *ClaudeMessagesResponse) IsToolUse() bool {
 	return r.StopReason == "tool_use"
 }
 
 // IsError checks if this is an error event
-func (e *ApiMessagesStreamEvent) IsError() bool {
+func (e *ClaudeMessagesStreamEvent) IsError() bool {
 	return e.Type == "error"
 }
 
 // IsMessageStart checks if this is a message start event
-func (e *ApiMessagesStreamEvent) IsMessageStart() bool {
+func (e *ClaudeMessagesStreamEvent) IsMessageStart() bool {
 	return e.Type == "message_start"
 }
 
 // IsMessageStop checks if this is a message stop event
-func (e *ApiMessagesStreamEvent) IsMessageStop() bool {
+func (e *ClaudeMessagesStreamEvent) IsMessageStop() bool {
 	return e.Type == "message_stop"
 }
 
 // IsContentBlockDelta checks if this is a content block delta event
-func (e *ApiMessagesStreamEvent) IsContentBlockDelta() bool {
+func (e *ClaudeMessagesStreamEvent) IsContentBlockDelta() bool {
 	return e.Type == "content_block_delta"
 }
